+++ simple-stack-example-scoping/src/main/kotlin/com/zhuinden/simplestackexamplescoping/WordListKey.kt
@@ -1,0 +1,11 @@
+package com.zhuinden.simplestackexamplescoping
+
+import kotlinx.android.parcel.Parcelize
+
+/**
+ * Created by Zhuinden on 2018.09.17.
+ */
+@Parcelize
+data class WordListKey(val placeholder: String = "") : BaseKey, WordScope {
+    override fun createFragment() = WordListFragment()
+}