+++ simple-stack-example-scoping/src/main/kotlin/com/zhuinden/simplestackexamplescoping/WordScope.kt
@@ -1,0 +1,12 @@
+package com.zhuinden.simplestackexamplescoping
+
+import com.zhuinden.simplestack.History
+import com.zhuinden.simplestack.ScopeKey
+
+interface WordScope : ScopeKey.Child {
+    companion object {
+        const val SCOPE_TAG = "WORD"
+    }
+
+    override fun getParentScopes(): MutableList<String> = History.of(SCOPE_TAG)
+}