+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+-Simple Stack ?.??.? (????-??-??)
+--------------------------------
+
+- UPDATE: Added `simple-stack-example-scoping` to showcase the usage of `ScopeKey.Child`, with Fragments and Navigator.
+
 -Simple Stack 1.13.0 (2018-09-10)
 --------------------------------
 - ADDED: Adds `ScopeKey.Child` interface, which allows the definition of *explicit* parent hierarchy of a given scope.